potwierdzenie usuniecia pracownika

+++ MotoFitAcademy/OpenDayApplication/Model/Managers/WorkersManager.cs
@@ -57,18 +57,28 @@ namespace OpenDayApplication.Model.Managers
     }
     public void DeleteWorker(Worker worker)
     {
-            try
+            MessageBoxResult result = MessageBox.Show("Czy na pewno chcesz usunąć pracownika?", "Potwierdzenie", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
             {
-                using (var dataContext = new MotoFitAcademyDataContext(Confiuration.GetSqlConnectionString()))
+                try
                 {
-                    dataContext.Workers.Attach(worker);
-                    dataContext.Workers.DeleteOnSubmit(worker);
-                    dataContext.SubmitChanges();
+                    using (var dataContext = new MotoFitAcademyDataContext(Confiuration.GetSqlConnectionString()))
+                    {
+                        dataContext.Workers.Attach(worker);
+                        dataContext.Workers.DeleteOnSubmit(worker);
+                        dataContext.SubmitChanges();
+                    }
                 }
-            } catch (System.Data.SqlClient.SqlException)
-            {
-                System.Windows.MessageBox.Show("Nie udało się usunąć pracownika.", "Błąd połączenia z bazą danych");
+                catch (System.Data.SqlClient.SqlException)
+                {
+                    System.Windows.MessageBox.Show("Nie udało się usunąć pracownika. Błąd połączenia z bazą danych lub pracownik jest przypisany do zajęć.");
+                }
             }
+            else
+            {
+                Application.Current.Shutdown();
+            }
+            
     }
     public void EditWorker(Worker worker)
     {
